feat: add `chain_config_errors` to orchestrator's `UpdateNodeResponse`
To notify the core node of partial failures when updating node configs in CLO,
we're adding a `chain_config_errors` field to the `UpdateNodeResponse` message.

When a partial failure occurs -- for instance, one or more node configurations
do not have a matching network in CLO -- we return the protobuf with the
response but the new field will contain a map of `chain_id => error
message` for each config that could not be processed.

Ticket: [DPA-1371](https://smartcontract-it.atlassian.net/browse/DPA-1371)