feat: add errors to orchestrator's UpdateNodeResponse (#50)
To notify the core node of partial failures when updating node configs in CLO,
we're adding an `errors` field to the `UpdateNodeResponse` message.

When a partial failure occurs -- for instance, one or more node configurations
do not have a matching network in CLO -- we return the protobuf with the
response but the `errors` field will contain one entry for each node config
that could not be processed.

Ticket: [DPA-1371](https://smartcontract-it.atlassian.net/browse/DPA-1371)