Use Go Versioning Within Generated Code
This commit proposes moving versioned proto messages into subdirectories without separating different versions of
generated Go code. In this way, `v1` generates code to a single go package. If `v2` gets released in the future,
the new protos would replace the existing Go code and the package would be versioned as `v2`.

*v1*
github.com/smartcontractkit/chainlink-protos/workflows/go

*v2*
github.com/smartcontractkit/chainlink-protos/workflows/go/v2

The benefit of this pattern would be that the Go package follows idiomatic Go versioning while still separating
the proto versions. A consequence of this pattern is that it might be more difficult for both `v1` and `v2` to be
used at the same time in dependent projects.